Fixes #11608: Port external node info to rpkg format
