--- external-node-information/.gitignore
@@ -1,10 +1,0 @@
-target*
-.project*
-.classpath*
-.settings*
-.scala_dependencies*
-*logback.xml
-.idea*
-*.iml
-.externalToolBuilders
-
