--- external-node-information/packaging/SOURCES/.dependencies
@@ -1,1 +1,0 @@
-JAVA:jdk
