--- external-node-information/packaging/SOURCES/settings-external.xml
@@ -1,5 +1,0 @@
-<settings xmlns="http://maven.apache.org/SETTINGS/1.0.0" 
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0 http://maven.apache.org/xsd/settings-1.0.0.xsd">
-	<localRepository>/tmp/m2-repository</localRepository>
-</settings>
