--- external-node-information/packaging/SOURCES/settings-internal.xml
@@ -1,16 +1,0 @@
-<settings xmlns="http://maven.apache.org/SETTINGS/1.0.0" 
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0 http://maven.apache.org/xsd/settings-1.0.0.xsd">
-	<mirrors>
-		<mirror>
-			<!--This sends everything else to /public -->
-			<id>nexus</id>
-			<mirrorOf>*</mirrorOf>
-			<url>http://nexus.normation.com/nexus/content/groups/public</url>
-		</mirror>
-	</mirrors>
-	<localRepository>/tmp/m2-repository</localRepository>
-	<activeProfiles>
-		<activeProfile>nexus</activeProfile>
-	</activeProfiles>
-</settings>
