--- external-node-information/packaging/debian/README
@@ -1,5 +1,0 @@
-The Debian Package rudder-plugin-external-node-information
-----------------------------------------------------------
-
-Comments regarding the Package
-
