--- external-node-information/packaging/debian/compat
@@ -1,1 +1,0 @@
-7
