--- external-node-information/packaging/debian/conffiles
@@ -1,1 +1,0 @@
-opt/rudder/etc/plugins/external-node-information.properties
