--- external-node-information/packaging/debian/copyright
@@ -1,27 +1,0 @@
-This package was debianized by:
-
-    Rudder packaging team <rudder-packaging@rudder-project.org> on Wed, 12 Nov 2014 15:39:00 +0100
-
-It was downloaded from:
-
-    http://www.rudder-project.org/
-
-Upstream Author(s):
-
-    Normation SAS
-
-Copyright:
-
-    Copyright (C) 2014 Normation SAS
-
-License:
-
-    GNU General Public Licence version 3
-
-On Debian systems, the complete text of the GNU General
-Public License can be found in `/usr/share/common-licenses/GPL'.
-
-The Debian packaging is:
-
-    Copyright (C) 2014, Normation SAS
-
