--- external-node-information/packaging/debian/dirs
@@ -1,2 +1,0 @@
-opt/rudder/jetty7/rudder-plugins
-opt/rudder/etc/plugins
