--- external-node-information/packaging/debian/postinst
@@ -1,50 +1,0 @@
-#!/bin/sh
-# postinst script for rudder-inventory-ldap
-#
-# see: dh_installdeb(1)
-
-set -e
-
-# summary of how this script can be called:
-#        * <postinst> `configure' <most-recently-configured-version>
-#        * <old-postinst> `abort-upgrade' <new version>
-#        * <conflictor's-postinst> `abort-remove' `in-favour' <package>
-#          <new-version>
-#        * <postinst> `abort-remove'
-#        * <deconfigured's-postinst> `abort-deconfigure' `in-favour'
-#          <failed-install-package> <version> `removing'
-#          <conflicting-package> <version>
-# for details, see http://www.debian.org/doc/debian-policy/ or
-# the debian-policy package
-
-
-case "$1" in
-    configure)
-      # Register the plugin
-      if [ -x /opt/rudder/bin/rudder-plugin ]; then
-        /opt/rudder/bin/rudder-plugin register external-node-information
-      else
-        echo "********************************************************************************"
-        echo "* Unable to register the plugin automatically. Please enable it                *"
-        echo "* manually by editing your web application container context                   *"
-        echo "* file, adding /opt/rudder/jetty7/rudder-plugins/external-node-information.jar *"
-        echo "* to the <extraClasspath> anchor.                                              *"
-        echo "********************************************************************************"
-      fi
-    ;;
-
-    abort-upgrade|abort-remove|abort-deconfigure)
-    ;;
-
-    *)
-        echo "postinst called with unknown argument \`$1'" >&2
-        exit 1
-    ;;
-esac
-
-# dh_installdeb will replace this with shell code automatically
-# generated by other debhelper scripts.
-
-#DEBHELPER#
-
-exit 0
