--- external-node-information/packaging/debian/postrm
@@ -1,51 +1,0 @@
-#!/bin/sh
-# postrm script for normation-openldap-framework
-#
-# see: dh_installdeb(1)
-
-set -e
-
-# summary of how this script can be called:
-#        * <postrm> `remove'
-#        * <postrm> `purge'
-#        * <old-postrm> `upgrade' <new-version>
-#        * <new-postrm> `failed-upgrade' <old-version>
-#        * <new-postrm> `abort-install'
-#        * <new-postrm> `abort-install' <old-version>
-#        * <new-postrm> `abort-upgrade' <old-version>
-#        * <disappearer's-postrm> `disappear' <overwriter>
-#          <overwriter-version>
-# for details, see http://www.debian.org/doc/debian-policy/ or
-# the debian-policy package
-
-
-case "$1" in
-    remove|upgrade|failed-upgrade|abort-install|abort-upgrade|disappear)
-    
-    ;;
-    purge)
-      # Unregister the plugin
-      if [ -x /opt/rudder/bin/rudder-plugin ]; then
-        /opt/rudder/bin/rudder-plugin unregister external-node-information
-      else
-        echo "***********************************************************************************"
-        echo "* Unable to unregister the plugin automatically. Please disable it                *"
-        echo "* manually by editing your web application container context                      *"
-        echo "* file, removing /opt/rudder/jetty7/rudder-plugins/external-node-information.jar  *"
-        echo "* from the <extraClasspath> anchor.                                               *"
-        echo "***********************************************************************************"
-      fi
-    ;;
-
-    *)
-        echo "postrm called with unknown argument \`$1'" >&2
-        exit 1
-    ;;
-esac
-
-# dh_installdeb will replace this with shell code automatically
-# generated by other debhelper scripts.
-
-#DEBHELPER#
-
-exit 0
