--- external-node-information/packaging/plugin_versions.sh
@@ -1,21 +1,0 @@
-#####################################################################################
-# Copyright 2013 Normation SAS
-####################################################################################
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, Version 3.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-#####################################################################################
-
-
-export PLUGIN_VERSION=${PLUGIN_VERSION:="1.1.0"}
-export PLUGIN_RELEASE=${PLUGIN_RELEASE:="rc1~git"}
