--- external-node-information/pom.xml
@@ -1,73 +1,0 @@
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.normation.plugins</groupId>
-  <artifactId>external-node-information</artifactId>
-  <packaging>jar</packaging>
-
-  <parent>
-    <groupId>com.normation</groupId>
-    <artifactId>parent-pom</artifactId>
-    <version>3.1.16-SNAPSHOT</version>
-  </parent>
-  
-  <version>-3.1-1.1</version>
-
-  <description>
-    This is a plugin that allows to disaply external text (HTML, raw text) information about
-    nodes in the node details page. 
-    The plugin allows to configure a list of report name, root directory, description. 
-    When on node details, a new "external reports" tab is displayed, with the list
-    of report available for the node. 
-  </description>
-  
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-assembly-plugin</artifactId>
-        <version>2.4</version>
-        <configuration>
-          <descriptors>
-            <descriptor>src/main/assembly/plugin-with-own-dependencies.xml</descriptor>
-          </descriptors>
-        </configuration>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>single</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-
-  <dependencies>
-    <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>servlet-api</artifactId>
-      <version>2.5</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.normation.rudder</groupId>
-      <artifactId>rudder-web</artifactId>
-      <version>${rudder-version}</version>
-      <!-- this is needed as rudder-web is a war - its classes are packaged with "-classes" appened -->
-      <classifier>classes</classifier>
-      <scope>provided</scope>
-    </dependency>
-    
-  </dependencies>
-
-  <repositories>
-    <repository>
-      <id>maven.normation.com</id>
-      <url>https://maven.normation.com/</url>
-      <layout>default</layout>
-      <snapshots><enabled>true</enabled></snapshots>
-    </repository>
-  </repositories>
-
-</project>
