--- external-node-information/src/main/assembly/plugin-with-own-dependencies.xml
@@ -1,39 +1,0 @@
-<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0" 
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0 http://maven.apache.org/xsd/assembly-1.1.0.xsd">
-  <!-- TODO: a jarjar format would be better -->
-  <id>plugin-with-own-dependencies</id>
-  <formats>
-    <format>jar</format>
-    <format>dir</format>
-  </formats>
-  <includeBaseDirectory>false</includeBaseDirectory>
-  
-  <dependencySets>
-    <dependencySet>
-      <outputDirectory>/</outputDirectory>
-      <useProjectArtifact>true</useProjectArtifact>
-      <unpack>true</unpack>
-      <scope>runtime</scope>
-      <useTransitiveDependencies>true</useTransitiveDependencies>
-      <excludes>
-        <!-- 
-          We have to exclude all dependencies from parent-pom.
-          That seems utterly wrong, and so we seems to have done
-          something bad. 
-          Perhaps these dependencies should be on a "common-dependency" project, 
-          and included on all submodule, even if it seems not really efficient. 
-        -->
-         <exclude>org.scala-lang:scala-library</exclude>
-         <exclude>joda-time:joda-time</exclude>
-         <exclude>net.liftweb:lift-common_2.11</exclude>
-         <exclude>org.slf4j:slf4j-api</exclude>
-         <exclude>ch.qos.logback:logback-core</exclude>
-         <exclude>ch.qos.logback:logback-classic</exclude>
-         <exclude>ca.mrvisser:sealerate_2.11</exclude>
-         <exclude>org.scala-lang*</exclude>
-      </excludes>
-    </dependencySet>
-  </dependencySets>
-
-</assembly>
