--- external-node-information/src/main/resources/external-node-information.properties
@@ -1,44 +1,0 @@
-#
-# 
-#
-
-#The name of the tab in node details for report. 
-#Must be non empty
-plugin.external-node-information.tab-name=External Reports
-
-
-#
-# Configure several reports. 
-# Value must be enclosed with "" only
-# if the string contains "@"
-#
-# The reports will appear sorted alpha-numerically
-# by key, so you can force order by prefixing with
-# a number.
-#
-plugin.external-node-information.reports {
- 
-  01_security= {
-    title=Security Report
-    description=This report display pen test results
-    dirname=/tmp/security
-    filename="@@node@@.html"
-    content-type=text/html
-  }
-  
-  02_monitoring {
-    title=Monitoring Report
-    description=Monitoring information about the node
-    dirname=/tmp/monitoring
-    filename="@@node@@.txt"
-    content-type=text/plain
-  }
-
-  03_compliance {
-    title=Third party compliance report
-    description=Other compliance reports
-    dirname=/tmp/compliance
-    filename="@@node@@.pdf"
-    content-type=application/pdf
-  }
-}
