--- external-node-information/src/main/resources/nodeExternalReport.html
@@ -1,48 +1,0 @@
-<lift:surround with="menu/menu-administration" at="content">
-
-  <div class="portlet">
-    <div class="portlet-content">
-      <div class="inner-portlet">
-        <h4 class="inner-portlet-header">External information for Nodes</h4>
-        <div class="portlet-content">
-          <div class="intro">
-            This plugin allows to add external, static documents in a new tab in node details. 
-            <br />
-            With the plugin, you configure one (or several) root directory where are located 
-            your documents, and if a node has one for it (based on a regex and its hostname or uuid), 
-            the document will be downloadable. 
-            <br />
-            You can reload that page to reload the configuration of reports from the configuration file.
-          </div>
-
-          <div class="lift:PluginInfo">
-            <br /> <br />
-
-            <ul>
-              <li>File used to configure the reports: <b><span id="reportsConfigFile">[report config file]</span></b></li>
-              <li>Name displayed for the new tab in node detail: <b><span id="tabTitle">[tab title]</span></b></li>
-            </ul>
-            
-            <br /> <br />
-
-            <div class="oneReportDetails">
-              <h3>
-                Documents with key "<span class="reportKey">[key]</span>"
-              </h3>
-              <br/>
-              <ul>
-                <li><em>Title:</em> <b><span class="reportTitle">[title]</span></b></li>
-                <li><em>Description: </em><b><span class="reportDescription">[description]</span></b></li>
-                <li><em>Root directory: </em><b><span class="reportRootDirectory">[root directory]</span></b></li>
-                <li><em>Content Type:</em> <b><span class="reportContentType">[content type]</span></b></li>
-                <li><em>Example of file name for node "my-node.local": </em><b><span class="reportFilename">[file name example]</span></b></li>
-              </ul>
-              <br />
-            </div>
-          </div>
-        </div>
-      </div>
-      <br />
-    </div>
-  </div> 
-</lift:surround>
