+++ node-external-reports/.gitignore
@@ -1,0 +1,10 @@
+target*
+.project*
+.classpath*
+.settings*
+.scala_dependencies*
+*logback.xml
+.idea*
+*.iml
+.externalToolBuilders
+.worksheet/
