+++ node-external-reports/Makefile
@@ -1,0 +1,13 @@
+#
+# make all will build node-external-reports.rpkg. 
+# make demo will build a license limited version of the plugin
+#
+
+include ../common-scala-plugin.mk
+
+plugin: all
+
+plugin-resources:
+	cp src/main/resources/node-external-reports.properties target/$(NAME)/
+
+
