+++ node-external-reports/build.conf
@@ -1,0 +1,26 @@
+#
+# This file defines the release information about the plugin like
+# its version and its ABI compability. 
+#
+# So version are not managed in pom.xml (safe for parent-pom version,
+# which can't be a parameter, and must be equals to rudder-branch here)
+#
+
+# Unique identifier of the plugin
+plugin-id=rudder-plugin-node-external-reports
+# Human readable name 
+plugin-name=View external reports on node details
+
+# Plugin version. It is build as follow: A.B-x.y(.z) with:
+# - A.B: Rudder major.minor
+# - x.y(.z): plugin major.minor.micro. Micro should be omitted. When omitted, z is assumed to be 0. 
+# For the build, we split the information between two properties, rudder branch and plugin version, 
+# which must be concaneted with "-" to build the plugin version. 
+rudder-branch=4.1
+plugin-branch=1.2
+plugin-version=${rudder-branch}-${plugin-branch}
+
+# Version of Rudder used to build the plugin. 
+# It defined the API/ABI used and it is important for binary compatibility
+#
+rudder-build-version=4.1.0
