+++ node-external-reports/packaging/metadata
@@ -1,0 +1,11 @@
+{
+  "type": "plugin",
+  "name": "${plugin-id}",
+  "version": "${plugin-version}",
+  "build-date": "${maven.build.timestamp}",
+  "build-commit": "${commit-id}",
+  "jar-files": [ "/opt/rudder/share/plugins/node-external-reports/node-external-reports.jar" ],
+  "content": {
+    "files.txz": "/opt/rudder/share/plugins"
+  }
+}
