+++ node-external-reports/packaging/postinst
@@ -1,0 +1,2 @@
+#!/bin/bash
+
