+++ node-external-reports/src/main/resources/node-external-reports.properties
@@ -1,0 +1,44 @@
+#
+# 
+#
+
+#The name of the tab in node details for report. 
+#Must be non empty
+plugin.node-external-reports.tab-name=External Reports
+
+
+#
+# Configure several reports. 
+# Value must be enclosed with "" only
+# if the string contains "@"
+#
+# The reports will appear sorted alpha-numerically
+# by key, so you can force order by prefixing with
+# a number.
+#
+plugin.node-external-reports.reports {
+ 
+  01_security= {
+    title=Security Report
+    description=This report display pen test results
+    dirname=/tmp/security
+    filename="@@node@@.html"
+    content-type=text/html
+  }
+  
+  02_monitoring {
+    title=Monitoring Report
+    description=Monitoring information about the node
+    dirname=/tmp/monitoring
+    filename="@@node@@.txt"
+    content-type=text/plain
+  }
+
+  03_compliance {
+    title=Third party compliance report
+    description=Other compliance reports
+    dirname=/tmp/compliance
+    filename="@@node@@.pdf"
+    content-type=application/pdf
+  }
+}
