+++ node-external-reports/src/main/scala/com/normation/plugins/nodeexternalreport/service/NodeExternalReportApi.scala
