+++ node-external-reports/src/main/scala/com/normation/plugins/nodeexternalreport/snippet/PluginInfo.scala
